requirements.txt
streamlit
openai
google-generativeai
moviepy
pydub
python-dotenv
stripe
requests
imageio-ffmpeg